Updated NewTransaction to use FormDrawer.
The `NewTransaction` component now uses `FormDrawer` to do most of the heavy lifting now. Further, it
was renamed to `TransactionFormDrawer`.
